chipsec: run CHIPSEC without the -i option
The CHIPSEC team has advised that running with the -i is not a good idea
as it may lead to invalid results. This option is used mainly for
debugging purposes when it is necessary to ignore the fact that CHIPSEC
does not recognize the platform on which it is running. In most cases
the platform will be known. Also, it is better to simply say that a
given platform is not supported rather than giving spurious results.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>