chipsec: bump to v1.3.1
Now that a new version of CHIPSEC is out, update our recipe accordingly.

Bumping to this version required a trivial update to one of the patches
that we carry.

Also, a new change in v1.3.1 buils the CHIPSEC Linux module based on the
architecture of the host system. This prevents cross-compilation. Thus,
revert such patch.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>