umip: umip_test2: Print the complete expected result of tests
We use a 64-bit type to store result of tests. However, depending on the
used operands, the result can be 16, 32 or 64-bit. Hence, in the expected
value, print both the part of the variable that is expected to change
and the part that is expected to remain unchanged. This can help to better
understand failures.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>