umip: definitions: add macros to print test results
These macros are useful when we know the test result and its expected
value. Rather than writing this pattern multiple times, simply wrap it
in a macro.

We need two kinds of macros as sometimes the test results fits in a
single variable while y some other cases is of type table_desc.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>