umip: umip_ldt_64: Add a warning about segfault
This test program creates a local descriptor table using as segment bases
the GS and FS registers. These are used also by glibc for per-thread
data. The test program has a yet unidentified bug in the manner in which
it handles such registers. While this is investigated, add a warning
about this situation.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>