umip: umip_test_defs: add a macro to initialize expected signals
Depending on the configuration of the test suites and the underlying
Linux kernel, different signals are expected, as per the following rules.

If there is hardware support for UMIP, the default behavior is:
 + In 64-bit processes, UMIP-protected instructions are not emulated.
   In consequence, executing such instructions will invariably cause
   a #GP(0) (provided that no exeptions of higher priority happened.
 + In 32-bit processes, only SGDT, SIDT and SMSW are emulated. Hence,
   executing such instructions should not cause a #GP(0), the Linux
   kernel will provide emulation for them. The instructions STR and
   SLDT are not emulated and the #GP(0) exception will be observed.

However, it is possible to patch the Linux kernel to provide emulation for
all the instructions in both 32-bit and 64-bit processes. If this is the
case no #GP(0) exception should be observed (other exceptions can still
occur). In such a case, the test programs can be built by #define'ing
EMULATE_ALL

A couple of macros are provided to deal with the configurations described
above.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>