umip: umip_test: inspect possible signals after running UMIP-protected instructions
After running an UMIP-protected instruction, check if any signal was
was sent. If this is the case, compare the signal number and signal
code with expected values.

Before running the test, initialize got_signal and got_sigcode. This
implementation requires to remove the static qualifier to got_signal
and got_sigcode.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>