umip: umip_test: do not exit upon receiving a signal
Instead, let each test case inspect and handle the signal as per the
requirements of specific test cases.

Since we need to resume execution after the signal is received, we need
to increment the instruction pointer. We are not sure of the size
of the instruction size. We asume that the worst case scenario is that
of having two prefix bytes, two opcode bytes, one ModRM byte, one SIB
byte and 4 displacement bytes. A NOP sled is added after the instruction
to ensure we safely resume execution in case we overestimate the size
of the instruction.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>