umip: umip_test2: allow each test case handle received signals
Instead of exiting when receiving a signal, let each individual test
case to handle the signal and compare it with expected values.

When resuming execution we need to increment the instruction pointer
to the next instructions. It is difficult to calculate the exact size
of the instruction. Instead, we assume the largest value possible: two
prefix bytes, two opcode bytes, one ModRM byte, one SIB byte, 4
displacement bytes. We append at the end a NOP sled to ensure we
continue execution safely.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>