umip: LDT tests: update test generators to optionally generate code for STR and SLDT
STR and SLDT are not emulated unless the kernel is patched for such
purpose. Hence, it is useful to update the test generators to only
create test code for such instructions if needed. A new command line
argument for the script, --emulate-all is used for this purpose; i.e.,
test code for STR and SLDT will be generated if such parameter is
given in the command line.

TODO: Much of the code in these generator is repeated and can be put
in a single library. This is a good example.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>