umip: build test programs for emulation of all UMIP-protected instructions
It is convenient to have versions of the test programs that test both the
required functionality of UMIP emulation (i.e., emulate only 32-bit
processes and only the instructions SIDT, SGDT and SMSW) and the complete
emulation: emulate all UMIP-protected instructions in both 32 and 64-bit
processes. Update our Makefile for this purpose.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>