umip: umip_utils: improve signal handling
Modify signal handling function to prevent or allow further test case
validation. If a signal was expected and none was received, the test case
should fails. Likewise, if a signal was expected but the signal number
and/or signal code does not match what was expected, the case fails.
However, if a signal was received and it matches the expected signal
the test case passes. Conversely, if a signal not expected but one
is received, the test case fails. In all the aforementioned cases there is
no need to inspect the instructions' result.

Only when a signal was not expect and none is received the instructions'
results are relevant.

The function inspect_signal is updated to reflect the described behavior.
Also, the return value is used to indicate whether further results
inspection is needed.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>