umip: umip_test_opnds: correct local variable access due to unbalanced %(r|e)sp
Assembly test code must handle the stack pointer carefull so that it has
the same value it has before entering the test code. This is important as
we handle several automatic variables. Some compilers reference local
variables as offsets from the base pointer; others do it as offsets from
the stack pointer.

This patch focuses on making sure the stack pointer is preserved. We need
not to worry about the case when the base pointer is the test register; the
test code already restores the test regiter after the test.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>