umip: umip_test_opnd: use a scratch register in memory tests
We were using %(r|e)ax register as a scratch register for register-indirect
addressing tests. The scratch register is used to copy the test result into
a local variable. However, the test register is restored at the end of the
test before saving the test result. This means that when %(r|e)ax is the
test register, its value is retored at the end of the test and we lose the
test result. Hence, add the ability to specify which register is used
as scratch.

When using the macro, we use mostly registers %(r|e)ax, %(r|e)cx and
%(r|e)dx as scratch registers. These registers do no need to be preserved
in the System V calling convention. Also, our function is simple enough
to not use these registers to operate.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>