umip: umip_test_opnds: correct local variable access due to unbalanced %(r|e)sp
Assembly test code must handle the stack pointer carefully so that it has
the same value it had before entering the test code. This is important as
we handle several automatic variables. Some compilers reference local
variables as offsets from the base pointer; others do it as offsets from
the stack pointer.

This patch focuses on making sure the stack pointer is preserved. We need
not to worry about the case when the base pointer is the test register; the
test code already restores the test regiter after the test.

When testing instruction that use registers as operands, two scratch
registers are needed. The first scratch register is used to initialize the
test register and to retrieve the test result. The second one is used
solely to restor the stack pointer. This is needed in case the register
under test is such register. After testing we must immediately restore it
in order to keep our stack operations as well as our references to local
variables correct.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>