umip: umip_test_opnds: print test results with the correct size
In order to have a readable test output, test results are prefixed
with a zeros so that they reflect the size of the data being printed.
However, in 32-bit builds, it does not make sense to print numbers
of 16 digits. Hence, only print as many as prepend zeros as needed.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>