umip: umip_test_opnds: print correct expected value
Tests for register operand incorrectly print the obtained value instead
of the expected value. The correct way of printing the expected value
is to AND the expected value with the mask of changed bytes. The
result is ORed with intit value ANDed with the mask of bytes that are
not expected to change.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>