umip: umip_exceptions: make page-fault tests check for SI_KERNEL code
Some of the instructions that UMIP protects are not always emulated
(i.e., str and sldt). In such cases, a SIGSEGV with SI_KERNEL code is
received instead of the page fault.

Update the tests for page-faults to take the expected signal number
and signal code as arguments to be used when inspecting the received
signal.

This also requires having two sets of signal number and signal code:
one for smsw, sgdt and sidt, for which SIGSEGV/SEG_MAPERR is always
expected; and another for sldt and sidt, for which SIGSEGV/SEG_MAPERR
or SIGSEGV/SI_KERNEL are expected, depending on whether str and sldt
are emulated.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>