umip: umip_utils: use exit_on_signal codes to determine test status
If a program needs to exit on signal, it is considered that the test
program failed and needs to terminate. However, there may be cases
in which receiving a signal is expected (e.g, when a test program
is built without -DEMULATE_ALL). In that case, exit_on_signal can
be used to determine if the test generating the signal passed or
failed.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>