umip: umip_ldt_64: set exit on signal as passed test case
When this program is run with EMULATE_ALL undefined, no emulation
for any of the UMIP-protected instructions is expected. Thus, it is
expected to receive SIGSEGV/SI_KERNEL; the test passs. On the other
hand, if EMULATE_ALL is defined, the test programs should not receive
any signal and should exit if one is received. In this case, the test
failed.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>