meta-luv/README: Document how to use git with a proxy
After updating LUV to poky/pyro, the environment variable PATH is
sanitized such that the normal native binary paths (/bin, /sbin, /usr/bin
and so forth) are removed. We now need to explicitly specify the path to
the git proxy command, something which was automatically done earlier.

We have documented how to do this in this patch.

Reviewed-by: Ricardo Neri <ricardo.neri-calderon at linux.intel.com>
Signed-off-by: Megha Dey <megha.dey@linux.intel.com>