qemu: Add runtime dependency on OVMF
OVMF needs to have a build time dependency on qemu but qemu only needs
to have a run time dependency on OVMF. Currently QEMU has a build time
dependency on ovmf and ovmf has a dependency on qemu. This manifests
itself as a circular dependency.

Hence, adding OVMF to the run-time dependency list of QEMU.

Reviewed-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>
Signed-off-by: Megha Dey <megha.dey@linux.intel.com>