luv-crash-handler: Introduce a new KEXEC_APPEND variable
Apart from the current kernel parameters, we pass some additional
parameters as well when we reboot using kexec after a panic. Hence, we
currently use both the --reuse-cmdline as well as the --append options.

It would be better if we introduce a variable(KEXEC_APPEND) which contains
all the kernel parameters we need on a reboot after panic so that we would
only have to use the --append option.

Reviewed-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>
Signed-off-by: Megha Dey <megha.dey@linux.intel.com>