luv-test-manager: separate runner loop and counting
Currently, we pipe the output of the runner loop to awk which counts the
total number of testsuites, unit tests etc.

Functionally, it wouldn't be incorrect to separate these two tasks i.e
(1) run the runner loop and store all the results to a file  and (2) use
awk on that file to do the counting. This separation will be useful when
we introduce test suites which will reboot while running. Also , this
makes the code more readable.

Reviewed-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>
Signed-off-by: Megha Dey <megha.dey@linux.intel.com>