luv-test-manager: add a -a option to tee
By default tee command overwrites a file. The tee command can be instructed
to append to a file using the option –a. This will be useful when we
introduce test suites which have to reboot. Instead of overwriting the
result file for the test suites which cause a reboot, it will append to it.
Also, if the file to be appended does not exist, this change is innocouous;
a new file will be created.

Reviewed-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>
Signed-off-by: Megha Dey <megha.dey@linux.intel.com>