pstore-test: Add pstore selftests recipe
pstore selftests include scripts which run test cases to check pstore
behavior. This is useful to avoid regressions of pstore.

Pstore is used across kernel crash, so these test cases are split
into three parts.

 - pstore_tests: check pstore behavior before crash
 - pstore_post_reboot_tests: check pstore behavior after crash and
   reboot
 - pstore_crash_test: cause kernel crash and reboot

The pstore_test and the pstore_post_reboot_tests are the actual scripts
for testing pstore. On the other hand, the pstore_crash_test causes a
kernel panic and reboot. In addition, there is a "common_tests" script
which includes utilities and test cases used commonly in these scripts.

When the pstore_crash_test is executed, it creates a file as a reboot
flag. The pstore_post_reboot_tests detects whether the file exists or not.
If the file doesn't exists, the test cases are skipped.

This addresses the following issue:
https://github.com/01org/luv-yocto/issues/35

Reviewed-by: Ricardo Neri <ricardo.neri-calderon at linux.intel.com>
Signed-off-by: Megha Dey <megha.dey@linux.intel.com>