luv-test-manager: Create a file the first time LUV boots.
This is a preparatory patch to introduce test suites which may reboot
while running. On the first boot, we create a file called first_boot.
This acts as a flag which tells if LUV is running for the first time.

We create the LUV result folder with the timestamp as suffix and store it
into a variable; should there be a reboot because of any test-suite, we
will restore the LUV results folder name from this stored variable.
Thus, we will have just one results folder and append the post reboot
results to it instead of having 2 or more directories each with only the
results of the current boot.

For network boots, we do not have persistent storage available. Hence,
tests which reboot will not work in the same way as a disk boot. We
would not be able to append to the result files as they will be lost on
a reboot.

Reviewed-by: Ricardo Neri <ricardo.neri-calderon at linux.intel.com>
Signed-off-by: Megha Dey <megha.dey@linux.intel.com>