luv-test-manager: Prepare LUV for tests suites which reboot
We have introduced a file (first_boot) which is created when LUV boots
for the first time. This file will eventually contain the list of
test-suites that have already completed.

If a test-suite reboots while running, it is not placed in the first_boot
file. On the reboot, we start executing from the test-suite which caused
the reboot i.e start where we left off. Since first-boot contains the
number of completed test-suites, it is used in awk to print the number of
total test-suites.

Reviewed-by: Ricardo Neri <ricardo.neri-calderon at linux.intel.com>
Signed-off-by: Megha Dey <megha.dey@linux.intel.com>