linux-yocto-efi-test: pstore selftest reboot flag
The pstore self-tests create the reboot flag (which tells us that we are
in the reboot kernel) in a directory which does not survive the reboot in
LUV. In the luv_test_manager recipe, the result partition is mounted
using a variable LUV_STORAGE at the /mnt/luv-storage directory. Hence,
creating the reboot flag here will ensure it is present even after the
reboot.

Reviewed-by: Ricardo Neri <ricardo.neri-calderon at linux.intel.com>
Signed-off-by: Megha Dey <megha.dey@linux.intel.com>