luv: Enable pstore self tests
Add the pstore-test binary to the core-image-efi-initramfs image.
Also, if the user passes the luv.pstore_test parameter in the cmdline,
there will be a crash and since we use the same commandline on reboot,
we would again experience a reboot and this saga would continue. Hence
if the user specifies the luv.pstore_test option, we need to remove it
from the reboot kernel's commandline.

Reviewed-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>
Signed-off-by: Megha Dey <megha.dey@linux.intel.com>