kernel-efi-warnings: customize the number of leading and trailing lines
When logging error conditions detected in the kernel buffer, the 3 leading
and 3 trailing lines surrounded the error are by default. Different errors
are logged differently. Thus, it makes sense to give flexibility to select
the number of leading and trailing lines are logged when an error pattern
is detected.

Two new parameters are added to the test_dmesg function. If not provided
default values are again 3 leading lines and 3 trailing lines.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>