kernel-efi-warnings: add support for skipped tests
Some of the errors or warnings logged in the kernel message buffer require
specific configurations to be selected at build time. Hence, if no such
configurations are present, erroneous conditions will not be logged and we
might falsely report the test as passed. Instead, add use a new pattern to
detect the needed kernel configuration. If not found, skip the test.

Not all the kernel configurations that we use to search for errors can be
detected in advance. Thus, this parameter is optional.

Reviewed-by: Megha Dey <megha.dey@linux.intel.com>
Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>
Signed-off-by: Megha Dey <megha.dey@linux.intel.com>