kernel-efi-warnings: customize the number of leading and trailing lines
When errors are found in the kernel message buffer, the 3 leading and 3
trailing lines the error message are also reported. However, different
kernel submodules might log errors differently. Thus, it makes sense to
have flexibility to select the number of leading and trailing lines are
logged when an error pattern is detected.

Two new parameters are added to the test_dmesg function to select the
number of leading and trailing lines. If not provided, default values are
3 leading lines and 3 trailing lines.

Reviewed-by: Megha Dey <megha.dey@linux.intel.com>
Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>
Signed-off-by: Megha Dey <megha.dey@linux.intel.com>