luv-test-manager: Remove NUM_SUITES variable
We had introduced a NUM_SUITES variable to track the number of test suites
run. This was obtained by counting the number of words in the first boot
file (which had the names of test-suites whose runner was executed).
However, if a testsuite was removed at runtime(eg. bits) or couldn't run
at all(eg. chipsec on pre production systems/qemu), this value would still
report it to have completed which is incorrect. It turns out that counting
the number of '[-]' symbols in the luv.results file (which was done
initially) gives us the correct number of test-suites run.

Tested-By: Naresh Bhat <naresh.bhat@linaro.org>
Reviewed-by: Naresh Bhat <naresh.bhat@linaro.org>
Reviewed-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>
Signed-off-by: Megha Dey <megha.dey@linux.intel.com>