luv-test-parser: Add continue_parsing field to schema
For any test suite that runs several times but for which only a single
report must be generated(For eg. tests that cause a reboot), a new
continue_parsing field in introduced in the 1.2 schema.

Increment the schema version to v1.2 while retaining backwards
compatibility with v1.1. This allows parsers to be upgraded piece-meal
to the new schema.

Tested-By: Naresh Bhat <naresh.bhat@linaro.org>
Reviewed-by: Naresh Bhat <naresh.bhat@linaro.org>
Reviewed-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>
Signed-off-by: Megha Dey <megha.dey@linux.intel.com>