meta-luv/README: Provide more detail on luv.pstore-test
The luv.pstore-test command line argument introduced a while back enables
pstore-self tests(specifically the tests that cause a reboot and post
reboot tests).

Documentation of this parameter is somewhat ambiguous. Thus, we provide
more detail what this parameter does and clearly state that adding it
will cause the system to reboot.

Tested-By: Naresh Bhat <naresh.bhat@linaro.org>
Reviewed-by: Naresh Bhat <naresh.bhat@linaro.org>
Reviewed-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>
Signed-off-by: Megha Dey <megha.dey@linux.intel.com>