luv-test-manager: Remove first_boot file
We have introduced a first_boot file (to tell LUV that it is the first time
it is booting and stores variables to be restored across a reboot). This
file still remains in the luv results folder after luv completes and
should be removed just before we unmount the luv partitions.

Tested-By: Naresh Bhat <naresh.bhat@linaro.org>
Reviewed-by: Naresh Bhat <naresh.bhat@linaro.org>
Reviewed-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>
Signed-off-by: Megha Dey <megha.dey@linux.intel.com>