umip_utils: add a cleanup function for exit on signal
There are cases in which some cleanup may be need to be done when exiting
from a signal. Our signal handler cannot know what kind of cleanup is
needed. Thus, simply call a function that users of the handler can
implement as needed.

The cleanup function is NULL by default. Thus, users of our handler does
not need to implement it if not needed.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>