umip_ldt_64: implement a cleanup function for the signal handler
The signal handler provides functionality to invoke a cleanup function
before exiting. In the case of this program, we must ensure that we
restore the FS/GS base addresses for glibc to work properly.

This cleanup the same that is needed when tests run successfully. Thus,
this same function is called once tests are complete.

Reported-by: Pengfei Xu <pengfei.xu@intel.com>
Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>