luv-test-manager: Create directories only once
The luv-test-manager creates various directories to store the logs, raw
and parsed results of test suites as well as the html file to display
results. With the introduction of tests which cause a reboot, these
operations happen twice which is unnecessariy and in some cases
undesired.

Hence, we make sure that these operations occur only the first luv boots
The various directory/file names are stored and can later be restored if
a reboot occurs as they are required later by the luv-test-manager.

Tested-By: Naresh Bhat <naresh.bhat@linaro.org>
Reviewed-by: Naresh Bhat <naresh.bhat@linaro.org>
Reviewed-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>
Signed-off-by: Megha Dey <megha.dey@linux.intel.com>