pstore-test: Update runner
Currently, if the command-line does not have the luv.pstore-test parameter,
the pstore_tests and the pstore_post_reboot_tests are run. If this
parameter is present, the tests are run in this order: pstore_tests,
pstore_crash_test, pstore_tests, pstore_post_reboot_tests.

Ideally, we would want only the pstore_tests to run without this parameter.
If the parameter is present, tests should run in this order: pstore_tests,
pstore_crash_test, pstore_post_reboot_tests.

Also, the prev_uuid file remains behind in the result partition of LUV.
Hence, this needs to be deleted once the test has completed.

Tested-By: Naresh Bhat <naresh.bhat@linaro.org>
Reviewed-by: Naresh Bhat <naresh.bhat@linaro.org>
Reviewed-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>
Signed-off-by: Megha Dey <megha.dey@linux.intel.com>