luv-test: Ensure luv.reboot and luv.poweroff work
After running all the test-suites, command line parameters luv.reboot
and luv.powerof were intended to cause reboot or poweroff after a
specific timeout.

It was observed that both these options were not working after shifting
to systemD. This was becasue we were running the sleep process in
background and systemd kills any such chilld processes running in the
background.

We ths introduce a new script and service file, which sleeps for a
specified timeout and then

Signed-off-by: Megha Dey <megha.dey@linux.intel.com>