luv-test: Ensure luv.reboot and luv.poweroff work
After running all the test-suites, command line parameters luv.reboot
and luv.powerof were intended to cause reboot or poweroff after a
specific timeout.

It was observed that both these options were not working after shifting
to systemD. This was because we were running the sleep process in
background and systemd kills any such child processes running in the
background.

We thus introduce a new script and service file, which sleeps for a
specified timeout and then issues the relavant command: reboot or
poweroff.

This addresses the following issue:
https://github.com/01org/luv-yocto/issues/120

Reviewed-By/Tested-By: Naresh Bhat <naresh.bhat@linaro.org>
Reviewed-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>
Signed-off-by: Megha Dey <megha.dey@linux.intel.com>