bits: add the autogen dependency
Recipes can now fail due to auto-detection of undeclared/missing
dependencies (particularly but not limited to native tools). This means
all build-time dependencies for every recipe must be explicitly declared,
or they won't be populated into the sysroot for the recipe.

Since bits depends on autogen, we need to add it.

Reviewed-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>
Tested-By: Naresh Bhat <naresh.bhat@linaro.org>
Signed-off-by: Megha Dey <megha.dey@linux.intel.com>