luv-xxxx-image: Unify common parts of luv-live-image and luv-netboot-image
We currently maintain two recipes for live images. luv-live-image.bb is
used to generate disk images. luv-netboot-image.bb is used to generate
an EFI application image that can be booted via network.

While there are details that differ among these recipes, some parameters
are similar. To improve readability and maintainability of the project,
it would be good to have all these similarities in a common
.inc file (named luv-image.inc). Some similarities include

1. License and license file checksum
2. INITRD_IMAGE_LIVE and related variables
3. Linux command line, which varies slightly.
4. Default netconsole and storage URL parameters
5. Setting the EFI_USEXORRISO variable
6. Splash image

This patch addresses the following issue:
https://github.com/01org/luv-yocto/issues/114

Cc: Naresh Bhat <naresh.bhat@linaro.org>
Cc: Megha Dey <megha.dey@linux.intel.com>
Cc: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>
Reviewed-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>
Signed-off-by: Sai Praneeth Prakhya <sai.praneeth.prakhya@intel.com>