systemd: Selectively start systemd-networkd-wait-online service
The systemd-networkd-wait-online service runs by default and waits for
all links it is aware of to be configured. For systems that do not
have network enabled or with more than one link to configure, this service
waits for 2 minutes before giving up. This causes an unwanted delay for a
LUV user who does not want to use any feature which requires network
capability.

Thus, we start the system-networkd-wait online service only if the user
wants to use the following features: telemetrics, netconsole or send
LUV results over network to a server.

This addresses the following issue:
https://github.com/01org/luv-yocto/issues/124

Reviewed-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>
Signed-off-by: Megha Dey <megha.dey@linux.intel.com>