AArch64: Build grub-efi v2.02 for AArch64
We are trying to build and use grub across x86/ARM64 architecture.
Earlier there was no support of AArch64 in grub-efi, Hence AArch64
was dependent on upstream grub git repository. The LUV project use
to pull and build grub git repository whenever we build the project
for AArch64.  But recently poky updated meta layer with grub v2.02
and LUV is rebased on top of poky. Hence we can use a common grub
across all the architecture. The luv-netboot-image also depends on
the _RDEPENDS variable.

Signed-off-by: Naresh Bhat <naresh.bhat@linaro.org>
Signed-off-by: Megha Dey <megha.dey@linux.intel.com>