pstore-test: delete prev_uuid only in post-reboot tests
pstore tests use a string pattern to validate the integrity of the logged
data after a crash. Just before causing a kernel crash, such string pattern
is saved in REBOOT_DIR/prev_uuid. However, this file is only present when
testing pstore with kernel crashes. If kernel crash tests are not executed,
this file is not present. In such a case, we see an ugly error in the
results of pstore:

      rm: cannot remove '/mnt/luv-storage/prev_uuid':
      No such file or director

Logic could be added to only delete this file after a kernel crash test
has been executed. However, in order to keep the test runner simple, use
-f modifier in rm. In this manner, the file will be deleted when present
or else rm will exit silently.

Cc: Megha Dey <megha.dey@intel.com>
Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>
Signed-off-by: Megha Dey <megha.dey@linux.intel.com>