fwts: acpi: madt: only inspect Local APIC/x2APIC/SAPIC tables if enabled
The Advanced Configuration and Power Interface (ACPI) specification
version 6.2 Errata A in sections 5.2.12.2, 5.2.12.10 and 5.2.12.12 states
that "when a processor is not present, the Processor Local APIC/x2APIC/
SAPIC information is either not reported or flagged as disabled."

This implies that it is possible for firmware implementations to include
Local APIC/SAPIC/x2APIC tables for non-existent processors provided that
they are disabled.

Thus, only test these tables if they have the Enable flag set. Otherwise,
skip the test. Also, print a message to record the fact that these tables
exist but are not tested.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>
Reported-by: Bin Song <binx.song@intel.com>
Signed-off-by: Megha Dey <megha.dey@linux.intel.com>