linux-yocto-efi-test: x86/mttr: Update only valid variable-range MTRRs
Both the Intel 64 and IA-32 Architectures Software Developer's Manual Vol.
3 in section 11.11 and the AMD Architecture Programmer's Manual Vol. 2 in
section 7.7.5 state that in a multi-processor system Memory-Type
Range Registers (MTRRs) must provide a type of memory caching for all the
physical memory in the system that is consistent across all the processors.

For the specific case of the variable-range MTRRs, this implies that their
values should be the same across all processors.

Variable-range MTRRs utilize a pair of registers, MTRRphysMask_MSR and
MTRRphysBase_MSR, to describe the base, range and type of cache used for a
memory region. Also, a Valid bit is included in MTRRphysMask_MSR to enable
or disable a variable-range MTRR pair.

Hence, the condition of consistency across processors described above
applies to variable-range MTRRs that are enabled. Disabled variable-range
MTTR pairs can be inconsistent as long as they remain disabled.

Thus, when a variable-range MTRR pair is disabled, it is not necessary to
update MTRRphysBase_MSR as it will not be used. MTRRphysMask_MSR must be
updated in all cases as it contains the Valid bit that enables or disables
the pair.

Also, we want to warn about inconsistent MTRRs settings from BIOS. BIOS
might have configured only variable-range MTRR pairs that are enabled.
Such optimization was recently implemented in Tianocore: 9c8c4478cfca
("UefiCpuPkg/MtrrLib: Skip Base MSR access when the pair is invalid")

Thus, report updates to inconsistent pairs only when the pair remains valid
after the update or the update resulted in a change of the enabled state
(i.e., the pair transitioned from enabled to disabled, or viceversa).

This fixes Github issue https://github.com/01org/luv-yocto/issues/127.

Cc: Ruiyu Ni <ruiyu.ni@intel.com>
Reported-by: Bin Song <binx.song@intel.com>
Reviewed-by: Ruiyu Ni <ruiyu.ni@intel.com>
Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>
Signed-off-by: Megha Dey <megha.dey@linux.intel.com>