luv-tests: Design/build LUV such that each testsuite is enabled or disabled independently.
With "noluv" parameter, LUV disables not one but all testsuites which is
not adequate when user wants to disable specific testsuite(s).

The testsuites are enabled by default. Disable a specific testsuite
if needed by just removing the string that represents the testsuite in
"LUV_TESTS" parameter in luv.cfg file.

Cc: Ricardo Neri <ricardo.neri@intel.com>
Reviewed-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>
Signed-off-by: Gayatri Kammela <gayatri.kammela@intel.com>
Signed-off-by: Megha Dey <megha.dey@linux.intel.com>