grub: Change default tsc calibration method to pmtimer on EFI systems
The following grub patch is needed because presently grub hangs on Intel
Soc systems which use 8254 timer to calibrate TSC. This patch makes
pmtimer default instead of 8254 timer. This patch has already been
upstreamed to grub but is not a part of grub 2.02 (present version of
grub used by LUV). So, let's carry this patch in LUV until it's
released in some newer grub version.

Cc: Naresh Bhat <naresh.bhat@linaro.org>
Cc: Megha Dey <megha.dey@linux.intel.com>
Cc: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>
Reviewed-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>
Signed-off-by: Sai Praneeth Prakhya <sai.praneeth.prakhya@intel.com>
Signed-off-by: Megha Dey <megha.dey@linux.intel.com>