luv-test: Do not print unnecessary messages to console
Some scripts used in LUV print their messages to stdout. With LUV
switching to /dev/ttyS0(/dev/ttyAMA0 for aarch) as the standard
output, we see a lot of such prints which are not required. Thus,
redirect these messages to /dev/null instead.

Because of changing the standard output, some messages get printed
twice. This patch ensures that all messages are printed just once.

Reviewed-By: Naresh Bhat <naresh.bhat@linaro.org>
Reviewed-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>
Signed-off-by: Megha Dey <megha.dey@linux.intel.com>