bits: skip network configuration when unavailable
The script that configures the netconsole in preboot environment, seems
to keep waiting for the network to come up, so it can create a socket.
This waiting not only keep BITS from running but, gives an impression
that the system has hung when it is not connected to network.

This issue was seen in Minnowboard max2 with firmware version
(MNW2MAX1.X64.0097.R01.1709211052). To fix this issue, socket is created
successfully only after validating the netconsole parameters and if
there is network connectivity. If netconsole parameters are not
configured, it skips configuring the netconsole feature in BITS.

This patch has more checks included for every function call, so as to
save time when one of the crucial checks like valid parameters, network
connectivity and socket creation is failed by exiting the script.

Reported-by: Brian Richardson <brian.richardson@intel.com>
Signed-off-by: Gayatri Kammela <gayatri.kammela@intel.com>
Signed-off-by: Megha Dey <megha.dey@linux.intel.com>