luv-test: Install icons for HTML test report
At build time, install in the bootable image the graphic files that will
be used in the HTML test report.

At run time, copy such icons in the directory that contains the parsed
output of the test suites. This seems an appropriate location as it
contains all the files needed to generate the HTML report.

Files are copied in the html_start() function. This function handles the
setup of the HTML report.

For now, start with the icons used to represent the level of severity of
a failed test case.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>