luv-test-parser: Show the level of severity of test cases
Add functionality to show the level of severity of an error. The severity
is a positive number. The severity is depicted by prepending the test case
name by a group of icons whose number is equal to the given severity.

It is up to test suites' parser to define the level of severity. Ideally,
we should be consistent.

A reasonable set of severity levels are 1 icon for low, 2 icons for
medium, and 3 for high. Thus, if the severity is a negative number or
greater than 3, a question mark icon is displayed.

Also, add a legend at the end of the report to help the user interpret the
severity. A new CSS style is added for legends.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>