luv-test-parser: Show the level of severity of test cases
Add functionality to show the level of severity of failed test case. The
severity is a positive number and is depicted by appending to the name of
the test case a distinctive icon whose saturation of red. More severe
failed test cases will have redder icons. If the severity is a negative
number or greater than 3, a question mark is displayed.

It is up to test suites' parser to define the level of severity. Ideally,
we should be consistent.

Lastly, add a legend at the end of the report to help the user interpret
the severity. A new CSS style is added to be used for legends.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>