luv-test-parser: Add new schemas which include test case severity
Add a variant for the schemas 1.0, 1.1 and 1.2 that include a level of
severity for the result of a given test case.

These variants include a new field immediately after the field that has
the number of failed test cases.

In theory, this new severity field could be used for any of the stata of
the test case (i.e., passed. aborted, or skipped). However, severity
makes sense only for failed test cases.

TODO: Adding three new pragmas seems excessive. Perhaps all the parsers
could be updated. The change is the same for all of them.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>